Major logic implemented
Now can create custom todos. They are pushed into localStorage,
but as of now there is no logic to put them on the page (yet).

Added option to remove todo and option to mark it as completed
(removes custom todos from storage too). As of now marking todo
as completed doesnt do anything major, but it is tracked.