Somewhat works
Now it is possible to actually store and display all custom todos
and deleting actually deletes them from the storage.
Marking as completed works too.

However it is poorly implemented and should be changed, but i
have no idea how to be honest. get and set do not work even
if i set the prototype.